Reference frames through game_frames table
GameFrame model will hold the user frames for a user.
So it is a 1-1 mapping with users and 1-N with frames.

Column `active` will indicate whose turn it is to play. We will use this
information later validate the input data from api. Currently it active column
is not used but it is consumed in Game class so we make the tests to pass.