Removes GameFrame middle man
Previous design was not so good with the introduced GameFrame model.
There is no need to have a separate model to keep the current active hame
slot so I moved the relations to:

Game 1 ---- * User 1----- * Frame

The current player is now also computed by the first minumum
completed frames. This can change if it is needed to be persisted
but I don't think it will add a value at this point except the
overhead to keep track of synching it (see the removed OnGameFrameComplete)
which was not so good designed and had problems with test references.